Add new module for spectrum energy grouping
